requirements.txt
torch
torchvision
ftfy
regex
tqdm
Pillow
numpy
pandas
gradio
open_clip_torch
transformers
huggingface_hub
git+https://github.com/openai/CLIP.git
